# apps/front/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Temperature Outfit API"
  },
  "host": "localhost:3000",
...
  "paths": {
...
    "/api/outfit/{outfitId}": {
      "get": {
        "tags": ["outfit"],
        "description": "내 옷차림 조회",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "outfitId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "cuid"
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "schema": {
              "$ref": "#/definitions/OutfitResponse"
            }
          },
          "404": {
            "description": "일치하는 옷차림이 없음"
          }
        }
      },
      "put": {
        "tags": ["outfit"],
        "description": "내 옷차림 변경",
        "produces": ["application/json"],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "outfitId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "cuid"
          },
          {
            "name": "date",
            "in": "formData",
            "type": "string",
            "format": "date"
          },
          {
            "name": "image",
            "in": "formData",
            "type": "file | null"
          },
          {
            "name": "productsId",
            "in": "formData",
            "type": "string",
            "description": "여러개를 ','로 구별한다. ex) id1,id2,id3"
          },
          {
            "name": "comment",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "rating",
            "in": "formData",
            "type": "integer",
            "minimum": 0,
            "maximum": 10
          }
        ],
        "responses": {
          "200": {
            "description": "변경 성공",
            "schema": {
              "$ref": "#/definitions/OutfitResponse"
            }
          },
          "400": {
            "description": "값 누락"
          },
          "404": {
            "description": "내 옷차림이 아님"
          }
        }
      },
      "delete": {
        "tags": ["outfit"],
        "description": "내 옷차림 삭제",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "outfitId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "cuid"
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "schema": {
              "$ref": "#/definitions/OutfitResponse"
            }
          },
          "400": {
            "description": "값 누락"
          },
          "404": {
            "description": "내 옷차림이 아님"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProductResponse": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "description": "[nullable] 요청한 page"
        },
        "lastPage": {
          "type": "integer",
          "description": "[nullable] 최대 요청 가능한 page"
        },
        "limit": {
          "type": "integer",
          "description": "[nullable] 요청한 limit"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductDetailResponse"
          }
        }
      }
    },
    "ProductDetailResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "cuid",
          "example": "cl5qjhkic000409mi4nmw7epr"
        },
        "name": {
          "type": "string"
        },
        "categoryId": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "OutfitResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "cuid",
          "example": "cl5qjhkic000409mi4nmw7epr"
        },
        "date": {
          "type": "string",
          "format": "date"
        },
        "locationId": {
          "type": "integer"
        },
        "weather": {
          "$ref": "#/definitions/WeatherResponse"
        },
        "imageUrl": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductResponse"
          }
        },
        "comment": {
          "type": "string"
        },
        "rating": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "maximum": 100
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "WeatherResponse": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date"
        },
        "locationId": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "$ref": "#/definitions/WeatherStatus"
        },
        "temperature": {
          "type": "string",
          "description": "예보일 때 현재온도, 기록일 때 평균온도"
        },
        "lowestTemperature": {
          "type": "string"
        },
        "highestTemperature": {
          "type": "string"
        },
        "isForecast": {
          "type": "boolean",
          "description": "true면 예보, false면 기록"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "WeatherStatus": {
      "type": "string",
      "enum": ["sun", "cloud", "rain", "snow"]
    },
...
  }
}